+++ libs/ofxPointer/src/PointerUtilities.cpp
@@ -36,7 +36,7 @@ namespace ofx {
 
 #if !defined(TARGET_OSX)
 
-unsigned long long PointerUtilities::getSystemDoublePressInterval()
+uint64_t PointerUtilities::getSystemDoublePressInterval()
 {
     int timeMS = 0;
 #if defined(TARGET_LINUX)
@@ -44,7 +44,7 @@ unsigned long long PointerUtilities::getSystemDoublePressInterval()
     timeMS = g_settings_get_int(mouseSettings, "double-click");
     g_object_unref(mouseSettings);
 #elif defined(TARGET_WIN32)
-    timeMS = static_cast<unsigned long long>(GetDoubleClickTime());
+    timeMS = static_cast<uint64_t>(GetDoubleClickTime());
 #else
     timeMS = DEFAULT_DOUBLE_TAP_INTERVAL;
 #endif
